requirements.txt
fastapi
uvicorn
requests
langchain>=0.2.0
langchain-community
tiktoken
python-dotenv
faiss-cpu